client/requirements.txt
google-auth
google-cloud-storage
paho-mqtt
pyjwt